Fixed all of the filter features
Searching for a property or university name will instantly return any results that match that criteria. Adjusting the price slider will remove any cards that have a max price of more than the current slider value. Sorting by Room Type or Bedroom Count will sort the cards in ascending order according to each value. Also added a logo to the website, and changed some font colors to make everything more uniform.
