+++ unest-app/src/components/SearchBar.jsx
@@ -1,32 +1,24 @@
-import React, { createContext, Component, useState } from 'react';
-import Filter from "./Filter.jsx";
-import SliderComponent from './SliderComponent.jsx';
-import { slider } from '@material-tailwind/react';
+import React from 'react';
+import { Slider } from '@mui/material';
+import { Radio } from "@material-tailwind/react";
 
-function SearchBar() {
-  const [inputValue, setInputValue] = useState("");
-  const [filterValue, setFilterValue] = useState("");
-  const [sliderValue, setSliderValue] = useState(2000); //max value a property can have
-
-  function handleChange(data) {
-    // Update the state when input value changes
-    setInputValue(data);
-  };
-
-  function handleFilter(data) {
-    setFilterValue(data);
+function SearchBar({ handleFilter, handleSlider, handleChange, handleClick, sliderValue }) {
+  
+  function handleButtons(param) {
+    handleFilter(param);
   }
 
-  function handleSlider(data) {
-    setSliderValue(data);
+  function handleBar(value) {
+    handleSlider(value); 
   }
 
-  function handleClick() {
-    // Access the input value when the button is clicked
-    alert('Input value is: ' + inputValue + '\nFilter value is:' + filterValue + '\nSlider value is: ' + sliderValue);
-  };
+  function handleText(value) {
+    handleChange(value);
+  }
 
-
+  function handleSearch() {
+    handleClick();
+  }
 
   return <>
     <div>
@@ -37,10 +29,26 @@ function SearchBar() {
             <svg xmlns="http://www.w3.org/2000/svg" className="h-6 w-6 opacity-30" fill="none" viewBox="0 0 24 24" stroke="currentColor">
               <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" />
             </svg>
-            <input className="bg-gray-100 outline-none" onChange={handleChange} type="text" placeholder="Property/University Name" />
+            <input className="bg-gray-100 outline-none" onChange={(e) => handleText(e.target.value)} type="text" placeholder="Property/University Name" />
           </div>
-          <Filter handleFilter={handleFilter} handleSlider={handleSlider} sliderValue={sliderValue}/>
-          <div onClick={handleClick} className="bg-red-400 py-3 px-5 text-white font-semibold rounded-lg hover:shadow-lg transition duration-3000 cursor-pointer">
+          <div className="flex-row">
+            Sort by:
+            <div className='space-x-4'>
+              <Radio onChange={(e) => handleButtons(e.target.value)} name="type" label="Room Type" value='Room Type' />
+              <Radio onChange={(e) => handleButtons(e.target.value)} name="type" label="Bedroom Count" value='Bedroom Count' />
+            </div>
+            <div>
+              Max Price: {sliderValue}
+              <Slider
+                min={300}
+                max={2000}
+                defaultValue={sliderValue}
+                valueLabelDisplay="off" 
+                onChange={(event) => handleBar(event.target.value)}
+              />
+            </div>
+          </div>
+          <div onClick={handleSearch} className="bg-red-400 py-3 px-5 text-gray-800 font-semibold rounded-lg hover:shadow-lg transition duration-3000 cursor-pointer">
             <div>Search</div>
           </div>
         </div>
